requirements.txt
pillow==9.4.0
torch==2.2.1
torchvision==0.17.1
scikit-image==0.19.3
fastapi==0.111.0